refactor(landing):components names to pascal
